Patching error in bidirectional relationship maintenance.
A student pointed out the following scenario: We create two empty orchestras o1 and o2 and an instrument (say a Trumpet) i. Now, o1.addBrassInstrument(i) and o1 has i while i has o1. Now, o2.addBrassInstrument(i), ond o2 has i while i has o2 AND o1 still has i.

To correct this, we have two options. (1) We need to disallow adding an instrument to an orchestra if it already has an orchestra. We could do this in a REQUIRES (and perhaps throw on IllegalArgumentException). I chose not to. (2) We need to be able to REMOVE an instrument from an orchestra. If we can remove instruments from orchestras as a public method, we must also be able to remove orchestras from instruments.

(Also patched for the two quite accurate warnings about redundant initializers in Orchestra given by IntelliJ.)